% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resvg.R
\name{resvg}
\alias{resvg}
\title{Render an SVG to a nativeRaster object}
\usage{
resvg(svg_text, zoom = 1)
}
\arguments{
\item{svg_text}{svg as single text string}

\item{zoom}{zoom factor. Default NULL means to use the natural SVG size.}
}
\description{
Render an SVG to a nativeRaster object
}
